+++ package.json
@@ -28,10 +28,12 @@
     "mongoose": "^6.0.5",
     "multer": "^1.4.3",
     "node": "^16.6.1",
-    "nodemon": "^2.0.13",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^6.1.0",
     "path": "^0.12.7"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.13"
   }
 }
