+++ views/partials/header.ejs
@@ -39,6 +39,10 @@
                         <li class="nav-item">
                           <a class="nav-link active" href="/experiment" data-toggle="tooltip" data-placement="top" title="Currently here" > <b>LABS </b></a>
                         </li>
+                        <li class="nav-item">
+                          <a class="nav-link"  href="/logout">Logout </a>
+
+                        </li>
                       </ul>
                     </div>
                   </div>
